LayoutTests/media/loopstart-limits-expected.txt

  
Test loopstart attribute parsing, limits, and defaults.

<video loopstart=1.5> - loopstart specified, should not change
EXPECTED (video.loopStart == '1.5') OK
EXPECTED (audio.loopStart == '1.5') OK

<video start=1.0> - loopstart not specified, should return start attr
EXPECTED (video.loopStart == '1') OK
EXPECTED (audio.loopStart == '1') OK

<video> - no loopstart and no start, default should be 0
EXPECTED (video.loopStart == '0') OK
EXPECTED (audio.loopStart == '0') OK

<video loopstart='s1.0'> - bogus loopstart and no start, default should be 0 (start default)
EXPECTED (video.loopStart == '0') OK
EXPECTED (audio.loopStart == '0') OK

<video loopstart='s1.0' start=1.0> - bogus loopstart, default should be start attr
EXPECTED (video.loopStart == '1') OK
EXPECTED (audio.loopStart == '1') OK
END OF TEST
